* adapted tourney service and repository to the new implementation

+++ SuperGol/src/main/java/repositories/protocols/TourniesRepositoryProtocol.java
@@ -9,7 +9,7 @@ public interface TourniesRepositoryProtocol {
 
     List<Tourney> all();
 
-    Tourney find(Integer id);
+    Tourney find(Double id);
 
-    Fixture rankingForATourney(Integer id);
+    Fixture rankingForATourney(Double id);
 }
